# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AuthenticateDto:
    properties:
      authorization_code:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Failed to parse user data
        type: string
      message:
        example: Something went wrong
        type: string
    type: object
...
  dto.ErrorResponse500:
    properties:
      error:
        example: Internal Server Error
        type: string
      message:
        example: An unexpected error occurred on the server
        type: string
    type: object
...
  dto.LineTokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      id_token:
        type: string
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
...
  dto.RefreshTokenDto:
    properties:
      refresh_token:
        type: string
    type: object
...
  dto.SendEmailRequest:
    properties:
      data:
        additionalProperties: true
        description: Placeholders and their values
        type: object
      email_category:
        description: Can be ID (UUID) or unique Title of the category
        type: string
      email_content_label:
        description: Label of the specific email content within the category
        type: string
      language:
        allOf:
        - $ref: '#/definitions/enums.PageLanguage'
        description: Language of the email content
        enum:
        - th
        - en
      to_recipient_emails:
        items:
          type: string
        type: array
    required:
    - data
    - email_category
    - email_content_label
    - language
    type: object
  dto.SuccessLinkResponse:
    properties:
      item:
        type: string
      message:
        example: Success
        type: string
    type: object
...
  dto.SuccessTokenResponse:
    properties:
      item:
        $ref: '#/definitions/dto.LineTokenResponse'
      message:
        example: Success
        type: string
    type: object
...
  enums.PageLanguage:
    enum:
    - th
    - en
    type: string
    x-enum-varnames:
    - PageLanguageTH
    - PageLanguageEN
...
info:
  contact: {}
  description: This API supports both Web and admin (CMS) operations, including authentication,
    page management.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /common/authenticate:
    post:
      consumes:
      - application/json
      description: Accepts authentication data and returns a LINE login URL for OAuth
        2.0 authentication.
      parameters:
      - description: Authentication payload
        in: body
        name: authenticate_data
        required: true
        schema:
          $ref: '#/definitions/dto.AuthenticateDto'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated login link
          schema:
            $ref: '#/definitions/dto.SuccessTokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Initiate LINE Authentication
      tags:
      - Common - Line Login
  /common/login-link:
    get:
      description: Returns a LINE OAuth 2.0 login URL for redirect-based authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessLinkResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Generate LINE Login Link
      tags:
      - Common - Line Login
  /common/refresh-token:
    post:
      consumes:
      - application/json
      description: Accepts a refresh token and returns new access and refresh tokens
        for LINE OAuth 2.0.
      parameters:
      - description: Refresh token payload
        in: body
        name: refresh_toke_data
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refresh
          schema:
            $ref: '#/definitions/dto.SuccessTokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Refresh LINE Authentication Token
      tags:
      - Common - Line Login
  /emails/send:
    post:
      consumes:
      - application/json
      description: Sends an email using a pre-configured template identified by category,
        label, and language.
      parameters:
      - description: Email Sending Request
        in: body
        name: emailRequest
        required: true
        schema:
          $ref: '#/definitions/dto.SendEmailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted - Email queued for sending
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation Error or Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Email category or content template not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error or Email Sending Failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Send Email
      tags:
      - Email Sending
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"